Enable weights from CIv2
