--- tests/nightly/single_card/functional_vovnet/test_tt_vovnet.py
@@ -1,1 +1,0 @@
-../../../../models/experimental/functional_vovnet/tests/pcc/test_tt_vovnet.py
