+++ tests/nightly/single_card/vovnet/test_tt_vovnet.py
@@ -1,0 +1,1 @@
+../../../../models/experimental/vovnet/tests/pcc/test_tt_vovnet.py
